Implemented the first task 'Search characters by name in the characters list'
I decorate the GoTAdapter by extension (I know that the best option for decorate is compose instance of extends) but I don't know yet how to do that with the views. I tryed to do it wrapping the view but it wasn't korking well, so I decide to do it extending.